Changes necessary to compile mohid water in single precision
Changes necessary to compile mohid water in single precision. However, is not possible to compile still in single precision due to inconsistencies in call in module hydrodynamic and water properties of ModuleTwoWay subroutines to be discussed with João Sobrinho.

+++ Software/MOHIDWater/ModuleLagrangianGlobal.F90
@@ -27121,7 +27121,7 @@ DoCatch: do while (associated(CurrentOrigin))
     
         !Arguments------------------------------------------------------------------    
         character(len=*)                            :: PropName, Units
-        real,       dimension(:),  pointer          :: Array1D
+        real(8),    dimension(:),  pointer          :: Array1D
         integer                                     :: DimID
